Separate Relationship from RelationshipMeta
