+++ Source/Theo/Relationship.swift
@@ -12,53 +12,7 @@ let RelationshipDataFromNodeKey = "fromNode"
 let RelationshipDataToNodeKey   = "toNode"
 let RelationshipDataTypeKey     = "type"
 
-let TheoRelationshipExtensionsKey: String = "extensions"
-let TheoRelationshipStartKey: String      = "start"
-let TheoRelationshipPropertyKey: String   = "property"
-let TheoRelationshipSelfKey: String       = "self"
-let TheoRelationshipPropertiesKey: String = "properties"
-let TheoRelationshipTypeKey: String       = "type"
-let TheoRelationshipEndKey: String        = "end"
-let TheoRelationshipDataKey: String       = "data"
-let TheoRelationshipMetaDataKey: String   = "metadata"
 
-public struct RelationshipMeta: CustomStringConvertible {
-
-    let extensions: [String: Any]
-    let start: String
-    let property: String
-    let relationship_self: String
-    let properties: String
-    let type: String //TODO: add custom function so it will return RelationshipType
-    let end: String
-    let data: [String: Any]
-
-    public let metadata: [String: Any]
-    
-    public func relationshipID() -> String {
-        
-        let pathComponents: Array<String> = self.relationship_self.components(separatedBy: "/")
-        
-        return pathComponents.last!
-    }
-    
-    init(dictionary: Dictionary<String, Any>!) {
-        
-        self.extensions         = dictionary[TheoRelationshipExtensionsKey] as! Dictionary
-        self.start              = dictionary[TheoRelationshipStartKey]      as! String
-        self.property           = dictionary[TheoRelationshipPropertyKey]   as! String
-        self.relationship_self  = dictionary[TheoRelationshipSelfKey]       as! String
-        self.properties         = dictionary[TheoRelationshipPropertiesKey] as! String
-        self.type               = dictionary[TheoRelationshipTypeKey]       as! String
-        self.end                = dictionary[TheoRelationshipEndKey]        as! String
-        self.data               = dictionary[TheoRelationshipDataKey]       as! Dictionary
-        self.metadata           = dictionary[TheoRelationshipMetaDataKey]   as! Dictionary
-    }
-    
-    public var description: String {
-        return "Extensions: \(self.extensions), start \(self.start), property \(self.property), self \(self.relationship_self), properties \(self.properties), type \(self.type), end \(self.end), data \(self.data), relationshipID \(self.relationshipID()), metadata \(self.metadata)"
-    }
-}
 
 public struct RelationshipDirection {
     
@@ -98,9 +52,10 @@ public struct Relationship {
     
     // MARK: Private Properties
 
-    public fileprivate (set) var relationshipMeta: RelationshipMeta?
     fileprivate (set) var relationshipCreateMeta: [String:Any] = [String:Any]()
     fileprivate (set) var relationshipData: [String:Any]       = [String:Any]()
+    
+    public let id: UInt64
 
     // MARK: Lazy Properties
     
@@ -152,7 +107,7 @@ public struct Relationship {
 
     // MARK: Constructors
 
-    /// Designated Initializer
+    /// Rest Initializer
     ///
     /// - parameter Dictionary<String,Any>: data
     /// - returns: Relationship
@@ -164,11 +119,20 @@ public struct Relationship {
         
         if let dictionaryData: [String:Any] = data {
 
-            self.relationshipMeta = RelationshipMeta(dictionary: dictionaryData)
-            
-            if let metaForRelationship = self.relationshipMeta {
-                self.relationshipData = metaForRelationship.data
+            if let relationshipData = dictionaryData["data"] as? [String:Any] {
+                self.relationshipData = relationshipData
             }
+            
+            if let relSelf = dictionaryData["self"] as? String,
+                let stringId = relSelf.components(separatedBy: "/").last {
+                self.id = UInt64(stringId) ?? 0
+            } else {
+                self.id = 0
+            }
+        }
+        
+        else {
+            self.id = 0
         }
     }
     
@@ -276,10 +240,6 @@ extension Relationship: CustomStringConvertible {
         
         for (key, value) in self.relationshipData {
             returnString += "\(key): \(value) "
-        }
-        
-        if let meta: RelationshipMeta = self.relationshipMeta {
-            returnString += meta.description
         }
         
         for (key, value) in self.relationshipCreateMeta {
